+++ crates/gmsol/src/pda.rs
@@ -275,7 +275,7 @@ pub fn find_gt_exchange_pda(
 pub fn find_price_feed_pda(
     store: &Pubkey,
     authority: &Pubkey,
-    index: u8,
+    index: u16,
     provider: PriceProviderKind,
     token: &Pubkey,
     store_program_id: &Pubkey,
@@ -285,7 +285,7 @@ pub fn find_price_feed_pda(
             PriceFeed::SEED,
             store.as_ref(),
             authority.as_ref(),
-            &[index],
+            &index.to_le_bytes(),
             &[u8::from(provider)],
             token.as_ref(),
         ],